OS/Bottle: better size handling
* Add size_type and npos definition
* Change copy method:
 - copy(const Bottle& alt, int first = 0, int len = -1);
 + copy(const Bottle& alt, size_type first = 0, size_type len = npos);
* Small refactor
* The signature of a few methods changed from size_t to size_type (but
  they are actually the same thing).

+++ src/libYARP_OS/src/Bottle.cpp
@@ -242,7 +242,7 @@ int Bottle::getSpecialization()
     return implementation->getSpecialization();
 }
 
-void Bottle::copy(const Bottle& alt, int first, int len)
+void Bottle::copy(const Bottle& alt, size_t first, size_t len)
 {
     implementation->edit();
     if (alt.isNull()) {
